Working version (still failures in unittest)
